requirements.txt
django
djangorestframework
psycopg2-binary
djangorestframework_simplejwt
pytest
pytest-django
load_dotenv
flake8
